+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/Player/EffectHidePlayerClothes.lua
@@ -61,4 +61,6 @@ function EffectHidePlayerClothes:OnStart()
 
     local str = string.format(ChaosLocalization.GetString("misc", "clothes_hidden"), #clothes)
     ChaosPlayer.SayLineByColor(player, str, ChaosPlayerChatColors.removedItem)
+
+    ChaosPlayer.ScheduleItemsHiddenHint(player)
 end